+++ tests/error/nightly/derives_for_structs_with_backtrace.rs
@@ -1,12 +1,12 @@
 #![allow(dead_code)]
 
-use std::any;
+use core::error::{request_ref, request_value};
 
 use super::*;
 
 #[test]
 fn unit() {
-    assert!(any::request_ref::<Backtrace>(&SimpleErr).is_none());
+    assert!(request_ref::<Backtrace>(&SimpleErr).is_none());
 }
 
 #[test]
@@ -17,7 +17,7 @@ fn named_implicit_no_backtrace() {
         field: i32,
     }
 
-    assert!(any::request_ref::<Backtrace>(&TestErr::default()).is_none());
+    assert!(request_ref::<Backtrace>(&TestErr::default()).is_none());
 }
 
 #[test]
@@ -35,7 +35,7 @@ fn named_implicit_backtrace_by_field_name() {
         backtrace: Backtrace::force_capture(),
         field: 0,
     };
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
+    assert!(request_ref::<Backtrace>(&err).is_some());
     assert_bt!(==, err);
 }
 
@@ -52,7 +52,7 @@ fn named_implicit_backtrace_by_field_type() {
         implicit_backtrace: Backtrace::force_capture(),
         field: 0,
     };
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
+    assert!(request_ref::<Backtrace>(&err).is_some());
     assert_bt!(==, err, implicit_backtrace);
 }
 
@@ -68,7 +68,7 @@ fn named_explicit_no_backtrace_by_field_name() {
 
     type MyBacktrace = Backtrace;
 
-    assert!(any::request_ref::<Backtrace>(&TestErr {
+    assert!(request_ref::<Backtrace>(&TestErr {
         backtrace: Backtrace::force_capture(),
         field: 0
     })
@@ -85,7 +85,7 @@ fn named_explicit_no_backtrace_by_field_type() {
         field: i32,
     }
 
-    assert!(any::request_ref::<Backtrace>(&TestErr {
+    assert!(request_ref::<Backtrace>(&TestErr {
         implicit_backtrace: Backtrace::force_capture(),
         field: 0
     })
@@ -108,7 +108,7 @@ fn named_explicit_backtrace() {
         explicit_backtrace: Backtrace::force_capture(),
         field: 0,
     };
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
+    assert!(request_ref::<Backtrace>(&err).is_some());
     assert_bt!(==, err, explicit_backtrace);
 }
 
@@ -125,7 +125,7 @@ fn named_explicit_no_backtrace_redundant() {
 
     type MyBacktrace = Backtrace;
 
-    assert!(any::request_ref::<Backtrace>(&TestErr {
+    assert!(request_ref::<Backtrace>(&TestErr {
         not_backtrace: Backtrace::force_capture(),
         field: 0
     })
@@ -148,7 +148,7 @@ fn named_explicit_backtrace_by_field_name_redundant() {
         backtrace: Backtrace::force_capture(),
         field: 0,
     };
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
+    assert!(request_ref::<Backtrace>(&err).is_some());
     assert_bt!(==, err);
 }
 
@@ -166,7 +166,7 @@ fn named_explicit_backtrace_by_field_type_redundant() {
         implicit_backtrace: Backtrace::force_capture(),
         field: 0,
     };
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
+    assert!(request_ref::<Backtrace>(&err).is_some());
     assert_bt!(==, err, implicit_backtrace);
 }
 
@@ -189,7 +189,7 @@ fn named_explicit_suppresses_implicit() {
         field: 0,
     };
 
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
+    assert!(request_ref::<Backtrace>(&err).is_some());
     assert_bt!(==, err, not_backtrace);
     assert_bt!(!=, err);
 }
@@ -210,8 +210,8 @@ fn named_implicit_no_backtrace_from_source() {
     };
 
     assert!(err.source().is_some());
-    assert!(any::request_ref::<Backtrace>(&err).is_none());
-    assert!(any::request_value::<i32>(&err).is_none());
+    assert!(request_ref::<Backtrace>(&err).is_none());
+    assert!(request_value::<i32>(&err).is_none());
 }
 
 #[test]
@@ -230,8 +230,8 @@ fn named_explicit_no_backtrace_from_source() {
     };
 
     assert!(err.source().is_some());
-    assert!(any::request_ref::<Backtrace>(&err).is_none());
-    assert!(any::request_value::<i32>(&err).is_none());
+    assert!(request_ref::<Backtrace>(&err).is_none());
+    assert!(request_value::<i32>(&err).is_none());
 }
 
 #[test]
@@ -250,9 +250,9 @@ fn named_explicit_backtrace_from_source() {
     };
 
     assert!(err.source().is_some());
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
-    assert_eq!(any::request_value::<i32>(&err), Some(42));
-    assert_bt!(==, err, any::request_ref::<Backtrace>(&err.err).unwrap());
+    assert!(request_ref::<Backtrace>(&err).is_some());
+    assert_eq!(request_value::<i32>(&err), Some(42));
+    assert_bt!(==, err, request_ref::<Backtrace>(&err.err).unwrap());
 }
 
 #[test]
@@ -273,10 +273,10 @@ fn named_implicit_different_source_and_backtrace() {
     };
 
     assert!(err.source().is_some());
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
-    assert_eq!(any::request_value::<i32>(&err), Some(42));
+    assert!(request_ref::<Backtrace>(&err).is_some());
+    assert_eq!(request_value::<i32>(&err), Some(42));
     assert_bt!(==, err, backtrace);
-    assert_bt!(!=, err, any::request_ref::<Backtrace>(&err.err).unwrap());
+    assert_bt!(!=, err, request_ref::<Backtrace>(&err.err).unwrap());
 }
 
 #[test]
@@ -298,10 +298,10 @@ fn named_explicit_different_source_and_backtrace() {
     };
 
     assert!(err.source().is_some());
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
-    assert_eq!(any::request_value::<i32>(&err), Some(42));
+    assert!(request_ref::<Backtrace>(&err).is_some());
+    assert_eq!(request_value::<i32>(&err), Some(42));
     assert_bt!(==, err, backtrace);
-    assert_bt!(!=, err, any::request_ref::<Backtrace>(&err.err).unwrap());
+    assert_bt!(!=, err, request_ref::<Backtrace>(&err.err).unwrap());
 }
 
 #[test]
@@ -310,7 +310,7 @@ fn unnamed_implicit_no_backtrace() {
     #[derive(Default, Debug, Error)]
     struct TestErr(i32, i32);
 
-    assert!(any::request_ref::<Backtrace>(&TestErr::default()).is_none());
+    assert!(request_ref::<Backtrace>(&TestErr::default()).is_none());
 }
 
 #[test]
@@ -320,7 +320,7 @@ fn unnamed_implicit_backtrace() {
     struct TestErr(Backtrace, i32, i32);
 
     let err = TestErr(Backtrace::force_capture(), 0, 0);
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
+    assert!(request_ref::<Backtrace>(&err).is_some());
     assert_bt!(==, err, .0);
 }
 
@@ -331,8 +331,7 @@ fn unnamed_explicit_no_backtrace() {
     struct TestErr(#[error(not(backtrace))] Backtrace, i32);
 
     assert!(
-        any::request_ref::<Backtrace>(&TestErr(Backtrace::force_capture(), 0))
-            .is_none()
+        request_ref::<Backtrace>(&TestErr(Backtrace::force_capture(), 0)).is_none()
     );
 }
 
@@ -345,7 +344,7 @@ fn unnamed_explicit_backtrace() {
     type MyBacktrace = Backtrace;
 
     let err = TestErr(Backtrace::force_capture(), 0, 0);
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
+    assert!(request_ref::<Backtrace>(&err).is_some());
     assert_bt!(==, err, .0);
 }
 
@@ -361,8 +360,7 @@ fn unnamed_explicit_no_backtrace_redundant() {
     type MyBacktrace = Backtrace;
 
     assert!(
-        any::request_ref::<Backtrace>(&TestErr(Backtrace::force_capture(), 0))
-            .is_none()
+        request_ref::<Backtrace>(&TestErr(Backtrace::force_capture(), 0)).is_none()
     );
 }
 
@@ -373,7 +371,7 @@ fn unnamed_explicit_backtrace_redundant() {
     struct TestErr(#[error(backtrace)] Backtrace, i32, i32);
 
     let err = TestErr(Backtrace::force_capture(), 0, 0);
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
+    assert!(request_ref::<Backtrace>(&err).is_some());
     assert_bt!(==, err, .0);
 }
 
@@ -391,7 +389,7 @@ fn unnamed_explicit_suppresses_implicit() {
         0,
     );
 
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
+    assert!(request_ref::<Backtrace>(&err).is_some());
     assert_bt!(==, err, .0);
     assert_bt!(!=, err, .1);
 }
@@ -407,8 +405,8 @@ fn unnamed_implicit_no_backtrace_from_source() {
     });
 
     assert!(err.source().is_some());
-    assert!(any::request_ref::<Backtrace>(&err).is_none());
-    assert!(any::request_value::<i32>(&err).is_none());
+    assert!(request_ref::<Backtrace>(&err).is_none());
+    assert!(request_value::<i32>(&err).is_none());
 }
 
 #[test]
@@ -422,8 +420,8 @@ fn unnamed_explicit_no_backtrace_from_source() {
     });
 
     assert!(err.source().is_some());
-    assert!(any::request_ref::<Backtrace>(&err).is_none());
-    assert!(any::request_value::<i32>(&err).is_none());
+    assert!(request_ref::<Backtrace>(&err).is_none());
+    assert!(request_value::<i32>(&err).is_none());
 }
 
 #[test]
@@ -437,9 +435,9 @@ fn unnamed_explicit_backtrace_from_source() {
     });
 
     assert!(err.source().is_some());
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
-    assert_eq!(any::request_value::<i32>(&err), Some(42));
-    assert_bt!(==, err, any::request_ref::<Backtrace>(&err.0).unwrap());
+    assert!(request_ref::<Backtrace>(&err).is_some());
+    assert_eq!(request_value::<i32>(&err), Some(42));
+    assert_bt!(==, err, request_ref::<Backtrace>(&err.0).unwrap());
 }
 
 #[test]
@@ -456,10 +454,10 @@ fn unnamed_implicit_different_source_and_backtrace() {
     );
 
     assert!(err.source().is_some());
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
-    assert_eq!(any::request_value::<i32>(&err), Some(42));
+    assert!(request_ref::<Backtrace>(&err).is_some());
+    assert_eq!(request_value::<i32>(&err), Some(42));
     assert_bt!(==, err, .1);
-    assert_bt!(!=, err, any::request_ref::<Backtrace>(&err.0).unwrap());
+    assert_bt!(!=, err, request_ref::<Backtrace>(&err.0).unwrap());
 }
 
 #[test]
@@ -476,8 +474,8 @@ fn unnamed_explicit_different_source_and_backtrace() {
     );
 
     assert!(err.source().is_some());
-    assert!(any::request_ref::<Backtrace>(&err).is_some());
-    assert_eq!(any::request_value::<i32>(&err), Some(42));
+    assert!(request_ref::<Backtrace>(&err).is_some());
+    assert_eq!(request_value::<i32>(&err), Some(42));
     assert_bt!(==, err, .1);
-    assert_bt!(!=, err, any::request_ref::<Backtrace>(&err.0).unwrap());
+    assert_bt!(!=, err, request_ref::<Backtrace>(&err.0).unwrap());
 }